bibliocli/requirements.txt
alabaster==0.7.12
autopep8==1.6.0
Babel==2.9.1
BiblioCLI==1.0
bleach==4.1.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
distlib==0.3.4
dnspython==2.2.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
filelock==3.4.2
identify==2.4.4
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.1
Jinja2==3.0.3
keyring==23.5.0
MarkupSafe==2.0.1
networkx==2.6.3
nodeenv==1.6.0
packaging==21.3
pkginfo==1.8.2
platformdirs==2.4.1
pockets==0.9.1
pre-commit==2.17.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pymongo==3.12.3
pyparsing==3.0.7
python-dotenv==0.19.2
python-jsonrpc-server==0.3.4
pytz==2021.3
PyYAML==5.4.1
readme-renderer==32.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==11.0.0
six==1.16.0
snooty-lextudio==1.11.6
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tqdm==4.62.3
twine==3.7.1
typing_extensions==4.0.1
ujson==1.35
urllib3==1.26.7
virtualenv==20.13.0
watchdog==1.0.2
webencodings==0.5.1
zipp==3.7.0
